feat(LinkedList): update private property names using ES2022 syntax
The private properties _head and _tail of the LinkedList class were updated to #head and tail, respectively, to conform to the new ES2022 syntax for declaring private properties using the # symbol.

This change was made to take advantage of the improved syntax and maintain compatibility with current and future JavaScript versions.

No associated issues or problems are known.